+++ CHANGELOG.md
@@ -10,12 +10,15 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Fix not logging OpenAPI queries when `log-query=main-query` is enabled by @steve-chavez in #4226
 - Fix not logging explain query when `log-query=main-query` is enabled by @steve-chavez in #4319
 - Fix not logging transaction variables and db-pre-request function when `log-query=main-query` is enabled by @steve-chavez in #3934
+- Fix not logging the JSON message to stderr on a `PGRST002` error by @laurenceisla in #4129
+- Fix reloading the Schema Cache unnecessarily on a `PGRST002` error by @laurenceisla in #4367
 
 ### Added
 
 - Bounded JWT cache using the SIEVE algorithm by @mkleczek in #4084
   + It now uses a fixed size cache instead of arbitrary sized cache.
 - Add `--ready` flag for postgrest healthcheck by @taimoorzaeem in #4239
+- Add a hint to the `PGRST002` error by @laurenceisla in #4367
 
 ### Changed